proto1/requirements.txt
bcrypt==3.1.6
cffi==1.12.3
Django==2.2.1
django-cuser==2017.3.16
django-extensions==2.1.7
django-model-utils==3.1.2
django-seed==0.1.9
django-username-email==2.2.4
Faker==1.0.5
graphviz==0.10.1
psycopg2==2.8.2
psycopg2-binary==2.8.2
pycparser==2.19
pyperclip==1.7.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
